requirements.txt
amqp==5.3.1
aniso8601==9.0.1
billiard==4.2.1
blinker==1.9.0
camelot-py==0.11.0
celery==5.4.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-RESTful==0.3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pluggy==1.5.0
prompt_toolkit==3.0.48
pycparser==2.22
pypdf==5.1.0
PyPDF2==3.0.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
six==1.16.0
tabula-py==2.10.0
tabulate==0.9.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
